% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{fetch_census_estimates}
\alias{fetch_census_estimates}
\title{Get census estimates for Maine counties}
\usage{
fetch_census_estimates(simplify = TRUE)
}
\arguments{
\item{simplify}{logical, if TRUE simplify into a familiar form}
}
\value{
table of county population, density, geoid and geometry
}
\description{
Get census estimates for Maine counties
}
