% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{fetch_covid19datahub}
\alias{fetch_covid19datahub}
\title{Fetch updated data from COVID19 data hub}
\usage{
fetch_covid19datahub(...)
}
\arguments{
\item{...}{further arguments for \code{\link[COVID19]{covid19}}}
}
\value{
tibble of COVID19 data hub data
}
\description{
Fetch updated data from COVID19 data hub
}
\seealso{
\url{https://covid19datahub.io/}
}
