% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{fetch_covid_cumulative}
\alias{fetch_covid_cumulative}
\title{Retrieve the current COVID-19 cumulative counts}
\usage{
fetch_covid_cumulative(
  pop = read_census(),
  save_file = TRUE,
  path = mecdc_path("cumcounts")
)
}
\arguments{
\item{pop}{sf MUTLIPOLYGON as per \code{\link{read_census}}}

\item{save_file}{logical, if TRUE save as CSV}

\item{path}{character, where to save the file (date timestamped CSV)}
}
\value{
table of COVID19 cumulative counts and densities with spatial geometry.
   If pop is \code{NULL} then just a simple table of cumualtive counts is returned.
}
\description{
Retrieve the current COVID-19 cumulative counts
}
