% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{mecdc_path}
\alias{mecdc_path}
\title{Retrieve a CDC path}
\usage{
mecdc_path(..., root = "/mnt/ecocast/projectdata/covid19/MeCDC")
}
\arguments{
\item{...}{path segements passed to file.path}

\item{root}{character, the root path to the CDC data directory}
}
\value{
a file/directory path (not tested for existence)
}
\description{
Retrieve a CDC path
}
