% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{merge_covid_census}
\alias{merge_covid_census}
\title{Merge COVID-19 and census data}
\usage{
merge_covid_census(x = read_covid_cumulative(pop = NULL), pop = read_census())
}
\arguments{
\item{x}{tibble, the COVID-19 cases}

\item{pop}{sf object, as per \code{\link{read_census}}}
}
\value{
the merged covid/census table as sf MULTIPOLYGON
}
\description{
Merge COVID-19 and census data
}
