% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{plot_cumulative}
\alias{plot_cumulative}
\title{Create a two panel plot of COVID-19, by count and by density}
\usage{
plot_cumulative(x = fetch_covid_cumulative())
}
\arguments{
\item{x}{sf object of merged COVID-19 and census data}
}
\value{
a 2-plot ggplot2 object
}
\description{
Create a two panel plot of COVID-19, by count and by density
}
