% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census.R
\name{read_census}
\alias{read_census}
\title{Read Maine population and density form US Census}
\usage{
read_census(
  filename = system.file("me_county_population.geojson", package = "MeCDC")
)
}
\arguments{
\item{filename}{character, the name of the file}
}
\value{
sf MULTIPOLYGIN object of county populations
}
\description{
Read Maine population and density form US Census
}
