% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{read_covid19datahub}
\alias{read_covid19datahub}
\title{Read the locally store covid19datahub table for Maine Counties}
\usage{
read_covid19datahub(
  filename = mecdc_path("cumcounts", "covid19datahub.csv"),
  add_mecdc = TRUE
)
}
\arguments{
\item{filename}{character, the name of the data file with COVID19datahub data}

\item{add_mecdc}{logical, if TRUE append any new Me CDC data}
}
\value{
tibble
}
\description{
Read the locally store covid19datahub table for Maine Counties
}
