% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{read_covid_cumulative}
\alias{read_covid_cumulative}
\title{Read the cumulative cases file for a given date}
\usage{
read_covid_cumulative(
  date = Sys.Date(),
  path = mecdc_path("cumcounts"),
  pop = read_census()
)
}
\arguments{
\item{date}{Date class, the day to retrieve}

\item{path}{the path to the data file}

\item{pop}{sf object, as per \code{\link{read_census}}}
}
\value{
the merged covid/census table as sf MULTIPOLYGON
}
\description{
Read the cumulative cases file for a given date
}
