% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid19.R
\name{update_covid19datahub}
\alias{update_covid19datahub}
\title{Update the locally store covid19datahub table for Maine Counties}
\usage{
update_covid19datahub(filename = mecdc_path("cumcounts", "covid19datahub.csv"))
}
\arguments{
\item{filename}{character, the name of the data file with COVID19datahub data}
}
\value{
tibble of COVID19 data hub data
}
\description{
Update the locally store covid19datahub table for Maine Counties
}
